requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic>=2.9.0
sqlalchemy>=2.0.35
openai>=1.51.0
python-multipart==0.0.10
python-dotenv==1.0.1
requests==2.32.0
